Contains changes to make layered code work
