+++ layered/storage/json.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 
 	"github.com/katzien/go-structure-examples/layered/models"
-	"github.com/nanobox-io/golang-scribble"
+	scribble "github.com/nanobox-io/golang-scribble"
 )
 
 const (
